Refactor naming conventions in NER datasets
