+++ flair/datasets/__init__.py
@@ -7,43 +7,43 @@ from .base import MongoDataset
 # Expose all sequence labeling datasets
 from .sequence_labeling import ColumnCorpus
 from .sequence_labeling import ColumnDataset
-from .sequence_labeling import AMHARIC_NER
-from .sequence_labeling import ANER_CORP
-from .sequence_labeling import AQMAR
-from .sequence_labeling import BIOFID
+from .sequence_labeling import NER_MASAKHANE_AMHARIC
+from .sequence_labeling import NER_ARABIC_ANER
+from .sequence_labeling import NER_ARABIC_AQMAR
+from .sequence_labeling import NER_GERMAN_BIOFID
 from .sequence_labeling import BIOSCOPE
-from .sequence_labeling import BUSINESS_HUN
+from .sequence_labeling import NER_HUNGARIAN
 from .sequence_labeling import CONLL_03
 from .sequence_labeling import CONLL_03_GERMAN
 from .sequence_labeling import CONLL_03_DUTCH
-from .sequence_labeling import ICELANDIC_NER
-from .sequence_labeling import WEBPAGES_NER
+from .sequence_labeling import NER_ICELANDIC
+from .sequence_labeling import NER_ENGLISH_WEBPAGES
 from .sequence_labeling import CONLL_03_SPANISH
 from .sequence_labeling import CONLL_2000
-from .sequence_labeling import DANE
-from .sequence_labeling import EUROPARL_NER_GERMAN
-from .sequence_labeling import GERMEVAL_14
-from .sequence_labeling import HAUSA_NER
-from .sequence_labeling import INSPEC
-from .sequence_labeling import JAPANESE_NER
-from .sequence_labeling import KINYARWANDA_NER
-from .sequence_labeling import LER_GERMAN
-from .sequence_labeling import LUGANDA_NER
-from .sequence_labeling import LUO_NER
-from .sequence_labeling import MIT_MOVIE_NER_SIMPLE
-from .sequence_labeling import MIT_MOVIE_NER_COMPLEX
-from .sequence_labeling import MIT_RESTAURANT_NER
-from .sequence_labeling import PERSON_NER
-from .sequence_labeling import NAIJA_PIDGIN_NER
+from .sequence_labeling import NER_DANISH_DANE
+from .sequence_labeling import NER_GERMAN_EUROPARL
+from .sequence_labeling import NER_GERMAN_GERMEVAL
+from .sequence_labeling import NER_MASAKHANE_HAUSA
+from .sequence_labeling import KEYPHRASE_INSPEC
+from .sequence_labeling import NER_JAPANESE
+from .sequence_labeling import NER_MASAKHANE_KINYARWANDA
+from .sequence_labeling import NER_GERMAN_LEGAL
+from .sequence_labeling import NER_MASAKHANE_LUGANDA
+from .sequence_labeling import NER_MASAKHANE_LUO
+from .sequence_labeling import NER_ENGLISH_MOVIE_SIMPLE
+from .sequence_labeling import NER_ENGLISH_MOVIE_COMPLEX
+from .sequence_labeling import NER_ENGLISH_RESTAURANT
+from .sequence_labeling import NER_ENGLISH_PERSON
+from .sequence_labeling import NER_MASAKHANE_NAIJA_PIDGIN
 from .sequence_labeling import NER_BASQUE
 from .sequence_labeling import NER_FINNISH
 from .sequence_labeling import NER_SWEDISH
-from .sequence_labeling import NER_YORUBA
-from .sequence_labeling import STACKOVERFLOW_NER
-from .sequence_labeling import SEMEVAL2010
-from .sequence_labeling import SEMEVAL2017
-from .sequence_labeling import TURKU_NER
-from .sequence_labeling import TWITTER_NER
+from .sequence_labeling import NER_MASAKHANE_YORUBA
+from .sequence_labeling import NER_ENGLISH_STACKOVERFLOW
+from .sequence_labeling import KEYPHRASE_SEMEVAL2010
+from .sequence_labeling import KEYPHRASE_SEMEVAL2017
+from .sequence_labeling import NER_TURKU
+from .sequence_labeling import NER_ENGLISH_TWITTER
 from .sequence_labeling import UP_CHINESE
 from .sequence_labeling import UP_ENGLISH
 from .sequence_labeling import UP_FINNISH
@@ -52,30 +52,24 @@ from .sequence_labeling import UP_GERMAN
 from .sequence_labeling import UP_ITALIAN
 from .sequence_labeling import UP_SPANISH
 from .sequence_labeling import UP_SPANISH_ANCORA
-from .sequence_labeling import WEIBO_NER
-from .sequence_labeling import WIKIANN
-from .sequence_labeling import WIKIGOLD_NER
-from .sequence_labeling import WIKINER_ENGLISH
-from .sequence_labeling import WIKINER_GERMAN
-from .sequence_labeling import WIKINER_DUTCH
-from .sequence_labeling import WIKINER_FRENCH
-from .sequence_labeling import WIKINER_ITALIAN
-from .sequence_labeling import WIKINER_SPANISH
-from .sequence_labeling import WIKINER_PORTUGUESE
-from .sequence_labeling import WIKINER_POLISH
-from .sequence_labeling import WIKINER_RUSSIAN
+from .sequence_labeling import NER_CHINESE_WEIBO
+from .sequence_labeling import NER_MULTI_WIKIANN
+from .sequence_labeling import NER_ENGLISH_WIKIGOLD
+from .sequence_labeling import NER_MULTI_WIKINER
 from .sequence_labeling import WNUT_17
 from .sequence_labeling import WSD_UFSAC
-from .sequence_labeling import WNUT_2020_NER
-from .sequence_labeling import XTREME
-from .sequence_labeling import REDDIT_EL_GOLD
+from .sequence_labeling import NER_ENGLISH_WNUT_2020
+from .sequence_labeling import NER_MULTI_XTREME
 from .sequence_labeling import NER_GERMAN_POLITICS
-from .sequence_labeling import EntityLinkingCorpus
-from .sequence_labeling import AQUAINT_EL
-from .sequence_labeling import OLD_GERMAN_EL
-from .sequence_labeling import AIDA_CONLL_EL
-from .sequence_labeling import IITB_EL
-from .sequence_labeling import TWEEKI_EL
+
+# Expose all entity linking datasets
+from .entity_linking import EntityLinkingCorpus
+from .entity_linking import NEL_REDDIT_EL
+from .entity_linking import NEL_AQUAINT
+from .entity_linking import NEL_OLD_GERMAN
+from .entity_linking import NEL_AIDA_CONLL
+from .entity_linking import NEL_IITB
+from .entity_linking import NEL_TWEEKI
 
 # Expose all document classification datasets
 from .document_classification import ClassificationCorpus
