+++ tests/test_datasets.py
@@ -102,7 +102,7 @@ def test_load_column_corpus_options(tasks_base_path):
 
 def test_load_germeval_data(tasks_base_path):
     # get training, test and dev data
-    corpus = flair.datasets.GERMEVAL_14(tasks_base_path)
+    corpus = flair.datasets.NER_GERMAN_GERMEVAL(tasks_base_path)
 
     assert len(corpus.train) == 2
     assert len(corpus.dev) == 1
@@ -143,7 +143,7 @@ def test_load_no_dev_data_explicit(tasks_base_path):
 
 def test_multi_corpus(tasks_base_path):
 
-    corpus_1 = flair.datasets.GERMEVAL_14(tasks_base_path)
+    corpus_1 = flair.datasets.NER_GERMAN_GERMEVAL(tasks_base_path)
 
     corpus_2 = flair.datasets.ColumnCorpus(tasks_base_path / "fashion", column_format={0: "text", 2: "ner"})
     # get two corpora as one
